NLPASS_SPELLCHECKER(Frontend and Spelling Checker Backend)/backend/requirements.txt
Flask
pyspellchecker
Flask-CORS
requests